requirements.txt
gTTS==2.5.1
moviepy==1.0.3
Pillow==10.4.0
